Move env.ts
